+++ src/v2.c
@@ -21,8 +21,8 @@ poly = fma(poly, r, D_M5_C2);
 poly = fma(poly, r, D_M5_C1);
 poly = fma(poly, r, D_M5_C0);
 
-double result =
-    ldexp(poly, (int32_t)n);
-return (float)result;
+float result =
+    ldexp((float)poly, (int32_t)n);
+return result;
 
 }
